Optimising the code to retrieve all the project countries
Retrieving all the project countries checking them from all the products is not the best, mostly when the project contains a lot of products. The risk is to get a timeout error from Commercetools due to an offset value above the valid threshold. It's better to retrieve the countries from the project information instead.